jobs/ml-ops/s3/data-store/requirements.txt
python-dotenv==1.0.0
boto3==1.29.3